# Use closure when initialize render